pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.121.2",
    "fastapi-mail>=1.5.8",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-orm>=1.2.10",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
